// src/lib/enums.ts
enum DifficultyNumber {
    basic = 0,
    advanced = 1,
    expert = 2,
    master = 3,
    remaster = 4,
}

enum Difficulty {
    Basic = 0,
    Advanced = 1,
    Expert = 2,
    Master = 3,
    ReMaster = 4,
}

enum ComboType {
    none = -1,
    FC = 0,
    FCp = 1,
    AP = 2,
    APp = 3,
}

enum SyncType {
    none = -1,
    FS = 0,
    FSp = 1,
    FDX = 2,
    FDXp = 3,
    SYNC = 4,
}

enum ChartType {
    STD = 0,
    DX = 1,
    UTAGE = 2,
}

enum TitleType {
    Normal = 0,
    Bronze = 1,
    Silver = 2,
    Gold = 3,
    Rainbow = 4,
}

export {
    DifficultyNumber,
    Difficulty,
    ComboType,
    SyncType,
    ChartType,
    TitleType,
};